+++ src/main/java/com/foodmate/backend/service/S3Deleter.java
@@ -1,4 +1,4 @@
-package com.foodmate.backend.util;
+package com.foodmate.backend.service;
 
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.SdkClientException;
