+++ src/main/java/com/foodmate/backend/service/S3Uploader.java
@@ -1,4 +1,4 @@
-package com.foodmate.backend.util;
+package com.foodmate.backend.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
